move tests for OrderDetails to OrderDetails/__tests__ folder
